Move old method of CtxContext to run_args.rs and make most of them private
